pyproject.toml
[tool.poetry]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = ["Pibomeister <edy7picazo@gmail.com>"]
readme = "README.md"
packages= [{ include = "graph" }]

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.7"
langgraph = "^0.2.53"
langchainhub = "^0.1.21"
langchain-community = "^0.3.7"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.3"
langchain-openai = "^0.2.9"
pypdf = "^5.1.0"
sentence-transformers = "^3.3.1"
langchain-huggingface = "^0.1.2"
rank-bm25 = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
